refactor go files
